ztunnel: split CA server from XDS server
Split the built-in CA server into its own package (pkg/ztunnel/ca) with
independent lifecycle management. The CA server retains the original
TCP+TLS transport on 127.0.0.1:15012 using bootstrap certificates,
unlike the XDS server which was switched to Unix Domain Sockets. The CA
server is only started when SPIRE is not enabled, controlled by the
enable-ztunnel-spire config option.

Changes:
- Create pkg/ztunnel/ca package with standalone CA server and gRPC endpoint
- CA server listens on TCP 127.0.0.1:15012 with TLS bootstrap credentials
- Update XDS server to only handle ADS (workload/service discovery)
- Add EnableSPIRE config option and enable-ztunnel-spire flag
- Conditionally mount cilium-ztunnel-secrets volume only when SPIRE is
  not enabled
- Skip cilium-ztunnel-secrets Secret creation when SPIRE is enabled
- Restore CA_ADDRESS Helm template to use caAddress value
  (default: https://localhost:15012)
- Add Helm template support for enable-ztunnel-spire based on
  authentication.mutual.spire.enabled
- Add unit tests for XDS server
