+++ install/kubernetes/cilium/templates/cilium-configmap.yaml
@@ -753,6 +753,9 @@ data:
     {{- end }}
   {{- else if eq .Values.encryption.type "ztunnel" }}
   enable-ztunnel: {{ .Values.encryption.enabled | quote }}
+  {{- if .Values.authentication.mutual.spire.enabled }}
+  enable-ztunnel-spire: "true"
+  {{- end }}
   {{- end }}
   {{- if .Values.encryption.nodeEncryption }}
   encrypt-node: {{ .Values.encryption.nodeEncryption | quote }}
