+++ install/kubernetes/cilium/templates/ztunnel/secret.yaml
@@ -1,4 +1,4 @@
-{{- if and .Values.encryption.enabled (eq .Values.encryption.type "ztunnel") }}
+{{- if and .Values.encryption.enabled (eq .Values.encryption.type "ztunnel") (not .Values.authentication.mutual.spire.enabled) }}
 {{- if .Values.encryption.ztunnel.secrets.bootstrapRootCert }}
 ---
 apiVersion: v1
