+++ pkg/ztunnel/xds/cell.go
@@ -9,7 +9,6 @@ import (
 	"github.com/cilium/hive/cell"
 	"github.com/cilium/statedb"
 
-	"github.com/cilium/cilium/pkg/endpointmanager"
 	"github.com/cilium/cilium/pkg/k8s/watchers"
 	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/ztunnel/config"
@@ -18,7 +17,7 @@ import (
 
 var Cell = cell.Module(
 	"ztunnel-xds",
-	"ztunnel certificate authority and control plane",
+	"ztunnel xDS control plane server",
 	cell.Provide(NewServer),
 	cell.Provide(func(x *Server) chan *EndpointEvent {
 		return x.endpointEventChan
@@ -31,7 +30,6 @@ type xdsServerParams struct {
 	Lifecycle              cell.Lifecycle
 	DB                     *statedb.DB
 	Logger                 *slog.Logger
-	EPManager              endpointmanager.EndpointManager
 	K8sWatcher             *watchers.K8sWatcher
 	Config                 config.Config
 	EnrolledNamespaceTable statedb.RWTable[*table.EnrolledNamespace]
@@ -45,7 +43,6 @@ func NewServer(params xdsServerParams) *Server {
 	server := newServer(
 		params.Logger,
 		params.DB,
-		params.EPManager,
 		params.K8sWatcher.GetK8sCiliumEndpointsWatcher(),
 		params.EnrolledNamespaceTable,
 		params.Config.XDSUnixAddr,
